feat: add onExceptionCaptured on errors during push notification init
- To implement the onExceptionCapture it has required to refactor
  pushNotification, spliting into 3 pieces:
  - the pushNotification itself
  - the backgroundListener
  - the pushNotificationHandler

  This arrangement avoids a circular dependency issue
  between pushNotification.js and reducers.js modules.
